Added: functions in index.js file divided to components
