+++ javascript/js-structure/js-structure_quiz-app/components/CardList/CardList.js
@@ -1,0 +1,13 @@
+import { cards } from "../../utils/Data.js";
+import { Card } from "../Card/Card.js";
+export function CardList() {
+  const cardList = document.createElement("section");
+  cardList.classList.add("card-list");
+
+  cards.forEach((card) => {
+    const cardElement = Card(card);
+    cardList.append(cardElement);
+  });
+
+  return cardList;
+}
