+++ javascript/js-structure/js-structure_quiz-app/index.html
@@ -4,9 +4,10 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <base href="/" />
     <title>Quiz-App</title>
-    <link rel="stylesheet" href="./styles.css" />
-    <script src="./index.js" defer></script>
+    <link rel="stylesheet" href="./styles.css" type="text/css" />
+    <script src="./index.js" type="module"></script>
   </head>
   <body>
     <div id="root"></div>
